mod: checking if libfuse2 is already installed

+++ src/_install_nvim.bash
@@ -6,8 +6,11 @@ LOG_FILE="$1"
 
 cd /tmp
 
-# installing nvim from the appimage also require libfuse2 to run
-sudo apt install libfuse2
+# libfuse2 is required to run appimages.
+# Checking if it's installed
+if ! dpkg -l libfuse2 > /dev/null 2>&1; then
+    sudo apt install libfuse2
+fi;
 
 # downloading the latest stable release of neovim
 wget --quiet https://github.com/neovim/neovim/releases/download/stable/nvim.appimage --output-document nvim
